requirements.txt
# requirements.txt

# Core FastAPI and Server
fastapi
uvicorn[standard]
python-dotenv

# Data and ML libraries
pandas
numpy
scikit-learn
xgboost

# External APIs and Services
sbrscrape
google-generativeai
requests
python-multipart

# Utilities
colorama
tqdm
toml

# --- Temporarily Disabled ---
# The tensorflow package requires complex C++ compilers on Windows.
# It has been commented out to allow the server to start.
# The application is designed to handle this by disabling the Neural Network model.
# tensorflow==2.14.0
